don't add notes to list if status suggests an error
this fixes #31 in the sense that no malformed note objects should be in the
list anymore. However this does not mean that notes whose fetch resulted in an
error are refetched. This would slow down the already slow fetching process
further more and most likely the important notes might be fetched without
problems. So this does not make a lot of sense at the moment. Retrying the
list fetch with `:Simplenote -l` in case the desired note is not there is a
much more sensible solution I think.