Change optional qty to optional since date for get_note_list()
get_note_list() was modified to support an optional quantity as an
argument to avoid having to retrieve the entire note list. This was
useful if you have a large number of notes. The intention was that you
could use it to retrieve the last 10 notes, etc.  However, the API does
not guarantee the order in which the note index is retrieved, therefore
limiting the list returned to 10 notes, for example, does not guarantee
you'll get the 10 most recent notes - it could be the 10 oldest.

The API does allow for a "since" parameter though, so it makes more
sense to use a date as an optional argument to get_note_list as a way of
limiting the number of notes retrieved.

The date has to be a string in YYYY-MM-DD format. If it fails to parse
it will be assumed to be a tag (due to the current way the optional
arguments to `-l` work).