Make the Simplenote list/index the only window on screen
Make the Simplenote list/index the only window on screen if no modified
buffers are open in the same tab page.

This avoids the "annoying" split that can occur, as mentioned in these
issues:

- https://github.com/mrtazz/simplenote.vim/issues/46
- https://github.com/mrtazz/simplenote.vim/issues/41

Using 'only' seems to be the best way to do it, rather than trying to
reuse the current buffer.