Update tests to include deleting a note
Need to make changes to simplenote.vim so buffers aren't closed when
trashing or deleting a note when running via Vader as it confuses things
there.

Also when listing the note index, need this to happen in the current
buffer when running via Vader.

Works, but probably still consider work in progress until I add last
test for open a note by key; decided not to add test for tagging as
requires user input.

References #17