Add a script to help automate release management
Shell script that, when passed an argument of the new version, will
change the version in plugin/simplenote.vim and update the CHANGELOG
with a list of all the commit messages since the last tagged release. It
then commits these files and tags the commit.

I have decided not to go as far as automatically pushing the release in
case manual tweaks are desired to the CHANGELOG.

Similarly for uploading a new version to vim.org; the other issue there
is that only the original author would be able to update vim.org anyway.

Fixes #24