Add test for tagging a note
Figured it out after a hint from Junegunn Choi, the author of Vader.vim.
Need to add a temporary mapping to the Simplenote -t function as that
way can pass text to it. Can then execute this in a normal mode Do
block.

Removed commment about no blank lines before first given block as this
was a bug that is now fixed in Vader.vim

References #17