Fix for evaluating whether tests are running. Affects UI
There's definitely some change that occured, I'm guessing from Vim 7.3
to 7.4 unless it's a change in Python? However, that doesn't excuse my
sloppy coding.

`vim.eval("exists('g:vader_file')")` evaluates to 0 or 1. Previously
using `not` in front of this would get a `true` or `false` as
appropriate, but now this no longer works.

I'm guessing before that `vim.eval(...)` returned a integer and now it's
a string? I should have explicity converted this to an integer from a
string anyway and tested for equality to a value as opposed to relying
on python to interpret 0 and 1 as False and True.

Fixes #52