Commit for ref only - Add tests for pinning / unpinning
This would be my preferred way to add tests for pinning and unpinning.
My very first attempts resulted in python errors when running under
Vader and seemingly it was unpinning that was causing the problem.
Adding sleep commands in allieviated the issues, but as it is now these
tests will still fail. The result will be:

A note to pin
unpinned

The pinning bit doesn't raise an error, it just doesn't work at all;
Nothing is appended to the buffer. Which is odd, because if pinning is
unsuccessful it should still append the entire note as a string.

Any way, been tearing my hair out for days trying to get this to work
and have given up. Wanted to commit this for reference, but I'll revert
it and follow-up with a commit that does work, albeit with a less ideal
test.