Re-order commands to avoid BufReadPost overriding intended filetype
The BufReadPost functionality introduced in
4c1d5b196ddab153b58bf0d00c373a0d47b864a2 could cause the filetype to be
overridden.

This mostly seemed to affect Gvim as opposed to Vim. BufReadPost causes
the file contents to be parses and then Vim tries to set the filetype
automatically based on the content. This meant markdown files would be
flagged as conf files (particularly if they started with a #). This bug
was due to the order that the commands were called in simplenote.vim:
I.e. BufReadPost being called after the attempts to set filetype based
on the Simplenote markdown flag or the global SimplenoteFileType
variable.

This commit basically changes the order of the commands to reset
anything BufReadPost might do.

I have used the word "this" too many times in this commit message.

Fixes #59
References #58