Quick fix for titles with single quotes causing error in list index
In the last commit I used Vim's strdisplaywidth to fix alignment of
titles/tags with multibyte characters in the list index. But I
hadn't considered that the first line of the note could contain a
single quote / apostrophe which meant the way I'd set up the
command would fail. I.e the following:

    vim.eval("strdisplaywidth('"+title+"')")

with a title of:

    A line containing a single quote ' like so

would evaluate as:

    vim.eval("strdisplaywidth('A line containing a single quote ' like so)")

And Vim would think there was an extra nonsensical argument to
strdisplaywidth.

So as a quick hacky fix I'm replacing all instances of single
quotes in the title when calculating the display width.

References #60