When deleting a note also remove it from the note index list
When a note is permanently deleted (:Simplenote -D) the corresponding
entry in the note index list buffer is also removed.

The corresponding entry in self.note_index is also removed so that you
can carry on opening notes correctly from the list index buffer.

I decided against just calling :Simplenote -l again because we'd have no
way of knowing whether it was called with a date limiter originally.
Also, manipulating the buffer directly is faster.

For the time being I'm not applying this to trashing a note as those
notes still do exist on Simplenote and so even though trashed notes are
not normally displayed in the index they can still be opened.

Fixes #62