Add "single window" mode to emulate behaviour of website
If a global variable of g:SimplenoteSingleWindow is set (to anything)
then simplenote.vim will attempt to emulate the behaviour of the
Simplenote website (and native applications) where only one note is
displayed at a time. The first note will be opened in a split window and
then that window reused for all notes (buffers) than are opened or
re-activated.

If g:SimplenoteSingleWindow is not set then the behaviour should remain
as it did before this commit: New notes opened will open in new split
windows and if an existing note (buffer) is opened from the list index
then that window is activated.

Unfortunately, because of the way Vim treats windows, there is no
guarantee this will always work as intended. One such instance would be
using the single window mode and then manually opening another window
in-between the list index and the "note". Vim will renumber the windows
and so new notes would open in this middle window. I don't know anyway
this can be prevented.

Another problem is that to achieve this functionality I have ended up
with quite a complicated mess of nested if statements. Hence why I have
also added comments to explain what is going on. I could do with
thinking about if it is possible to simplify the logic.

References #70