1st attempt at merging content when local and remote have been updated
Although this isn't exactly what the issue asked for, ultimately this
approach is the right fix. The whole point of Simplenote, afterall, is
that it makes merging changes to notes effortless. So rather than
reloading the remote into the current buffer, we should merged in the
changes (Simplenote does that automatically).

Initial (quick) testing shows that this worked. I tried two different
vim sessions and successfully merged content across them. I also tried
with one vim session and making changes on the Simplenote website.

Committing this on a separate branch for now as it needs more testing,
should also have Vader tests written for it (how??!) and also, from when
I wrote snose (https://github.com/atomicules/snose) I definitely had
issues when sending the version and not sending the modifydate. This
might not be an issue now as simplenote.py should include that anyway if
Simplenote.vim hasn't, but I could do with double-checking.

References #73