Add commands to allow previous versions of notes to be retrieved.
:Simplenote -v

Will display the version of the currently open note

:Simplenote -v X

Where X is an integer, will open that version of the note if it is
available into the buffer. If the current note/buffer is modified it
will prevent the user from displaying a previous version (to avoid
accidentally losing current changes) and will print an error message to
this effect. If it isn't available it prints an error message.  If you
want to restore this version just issue
`:Simplenote -u`

:Simplenote -V

Will open the most recent version of the current note into the buffer.
This is intended to be used with `:Simplenote -v` when you are browsing
through previous versions and you want to get back to the latest.

References #78