Use first line of note as note title, with key appended
Main thing to note here is that have to be careful with buffer names.
Vim will happily create buffer names containing characters such as
square brackets, but won't switch to them without escaping them,
therefore the safest thing to do is just display the alphanumeric
characters from the first line of the note.

Not closing out yet as wondering if it would be possible to remove the
key from the title somehow and store it in an index mapped to the buffer
number.

References #75