First attempt at removing note keys from buffer titles
A potential improvement on the previous commit
ed7cd65b7d9006ba9b1d002486c4efdceba755a3 which prepended the first line
of each note to the note key for the buffer title. That got me to
thinking about removing the need to have the note key in the buffer
title at all.

This works, but could be tidied more.

Uses another internal dictionary to map buffer numbers to note keys/ids
and takes advantage of the fact that, as far as I am aware, vim never
re-uses buffer numbers. The dictionary also has to be used in reverse,
but that seems fair enough - more a limitation of Python (or any
language in general?) that there is no good way of doing one to one
lookups in either direction.

References #75