More tweaks to retrieving versions of notes
- Don't let retrieving latest version overwrite modified buffer
- When retrieving another version set the buffer to unmodified

These two are somewhat in conflict, but I don't want someone's local
unsaved changes to be overwritten (even if they can get them back with
undo), but also, when browsing through versions you don't want the
buffer state to be modified as then it'd prompt you to save the
changes.

Had to re-work the "reset version" test a bit and the code in general so
the tests still work - because the note key is no longer in the title
the tests wouldn't run, so there needs to be an option in the code for
still having the title as the key when running tests.

This is a case of code ending up more messy and complicated just for the
sake of having tests.

References #78