Add note key back into title to ensure unique buffer names
Notes with the same first line would result in the same buffer name
which Vim doesn't allow. Therefore we need to add the key back into the
title / buffer name to ensure we always have something unique.

One plus side of this is that it means we can use the same title setup
for normal use and while testing via Vader. The downside though is
extracting the key from the title (for Vader) is more tricky.

Fixes #75

For sure this time.

I hope.