Bug fix - Don't overwrite buffers mods if re-opening from the list index
Previously, if you opened a note from the index and made some
modifications, but didn't save them and then went back to the list
index and pressed <Enter> on that note again it would re-open it (in the
same buffer) wiping out the local modifications.

This modification simply checks whether a buffer already exists for that
note and if it does it skips most of `display_note_in_scratch_buffer`
since it will have already been done.

Note: I'm trying to put together a test for this before I tag this as a
release, but currently struggling to get the test to work. I'll persist
a little longer before giving up.