Add test for previous commit 83932cb5fd910f2cad1b88a3b0dc6a2f3ec32ac0
Test to make sure that re-opening a note that is modified locally does
not overwrite with the remote content. I.e. so unsaved local changes
aren't lost.

I was trying to produce a test that replicated opening a note from the
list index, modifying the note, opening another note from the list index
and then re-opening the original, but I couldn't figure out getting this
to work.

So instead this test uses `:Simplenote -o` since that uses the same code
as opening a note from the list index.