Add Python 2 compatibility back in.
Fortunately a lot of Python 3 requirements work fine in Python 2:

- Using list on map
- Using items instead of iteritems
- Using brackets on print statements

However, the uppercase Q on "from Queue" is important under Python 2 so
have to test the version for this.

This is hopefully now done. Tests pass under Vader for Python 2 - not
sure if I can use Vader for Python 3, will just have to manually check
that.

Not merging in yet as want to give it a final review and a bit of a test
drive.

References #71