When using vertical split try to use correct width for note list index
Before, the entire width of the window was used and as soon as one note
was opened the tags and date, etc would be invisible. This is inspired
by the work of @sanchon in implementing the SimplenoteListSize setting.

This doesn't account for the Vim configuration setting of numberwidth,
but I'm not sure if it should.

References #87