Re-work sorting of notes in list index
This somewhat goes all the back to commit:
729f1516a514a17f502078eca40ff93ec8e75f80 where I somewhat lazily made
Python 3 compatible by using the `cmp_to_key` function.

Finally getting around to looking into #97 I realised the only way to
fix that issue was moving the sorting to an iterative based approach
where, say if we are sorting on "pinned, tags, modifydate", we first sort
the notes on modifydate, then go through them again and sort on tags and
finally on the primary key of pinned.

The `cmp_to_key` based function can't do this as it's only really doing
a single iteration through the notes, but trying to check all the
criteria (by iterating through the criteria) at once. This is why we
ended up with little groups within tags sorted by modified date.

Anyway, as a result of all this it's actually a lot simpler.

- Sort by tag is purely on existence of tags now (no alpha sort or
  number of tags sort)
- We are sorting tags within a note by alpha for display purposes; This
  should be moved to Simplenote.py

I've probably not explained any of that well.

Fixes: #97