Quick Simperium fixes
I am pushed for time. As with the simplenote.py changes for Simperium
I'm just trying to get something out that _works_ and then I can work on
making the code better.

It'd be easy if Python 2 just went away though.

- Use simplenote.py v2.0.0 (Simperium API)
- Make one test change; Haven't really looked into why it's now True
instead of 1. I.e. don't know if that's a Simperium change or something
I've not got quite right in the simplenote.py changes
- Handle the one Unicode / UTF-8 breakage I've found so far.