Try to keep list index width when vertical and using default width
Previously it picked 50%, which is all good until we started refreshing
the list index as it just halves the window again. Err.. that's stupid.
So instead only do that once, when the list index is first displayed.
Then keep that value.

Not going to be perfect, but if it ever goes screwy (I'm think ssh from
a terminal of a different size - maybe?) then at least the list index is
cache so it's easy to quit and start again.

Doesn't affect folk with a set list size or those still using the old
horizontal layout.