Cache authorisation token
Ummed and erred and decided to just write the token out to the note
index cache rather than another file. I am chmod'ing that file to be a
bit more secure.

Since that cache could already exist without that token, then check for
existence of that key first.

This is pretty simplistic, but I think better than the route I was going
down here:

https://github.com/mrtazz/simplenote.vim/commits/allow-token-instead-of-user-pass

where I was making it so you can provide a token instead of a password.
This way, worst comes to worst, people can edit the note index cache and
add a token or remove one, etc.

There is going to have to be a follow-up patch version release once I
make changes to simplenote.py to handle an invalid token. At the minute
that is not handled gracefully./

References: https://github.com/mrtazz/simplenote.vim/issues/103